EL_PulledAllUpdatesFromPreviousRepository

+++ Tools/Cosmology/ModeFunctions.h
@@ -13,10 +13,10 @@
 
 using namespace std;
 
-typedef double (*modefunction)(int, Parameters &);
+typedef double (*modefunction)(double, Parameters &);
 
-static double plainFunction(int a, Parameters &plainParameters){
-    return (double) a;
+static double plainFunction(double a, Parameters &plainParameters){
+    return a;
 }
 
 struct method{
